# frontend/landings/agivys/landing/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AgiVysSystem API",
    "description": "API para gestão do ecossistema AgiVysSystem",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://joederblanca.com.br/agivys-api"
    }
  ],
  "paths": {
    "/api/v1/authentication/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Autenticação de Usuário (Login)",
        "description": "Valida as credenciais do usuário e, se corretas, retorna um Token JWT junto com os dados essenciais do perfil e permissões.\r\n\r\n**Detalhes Importantes:**\r\n- O token gerado possui validade de **4 horas**.\r\n- O token deve ser enviado no cabeçalho `Authorization: Bearer {token}` em requisições subsequentes.\r\n- Além do token, a API também pode gerar um cookie `MedNext_Menu` (HttpOnly) contendo as permissões.",
        "requestBody": {
          "description": "Credenciais de acesso (E-mail e Senha).",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso. Retorna o Token e dados do usuário.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "E-mail ou senha incorretos."
          },
          "500": {
            "description": "Erro interno do servidor ao tentar autenticar."
          }
        }
      }
    },
    "/api/v1/authentication/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Encerramento de Sessão (Logout)",
        "description": "Realiza o logout do usuário limpando os cookies de sessão, em especial o cookie `MedNext_Menu`.\r\nObserve que, como o JWT é stateless, o cliente também deve descartar o token do lado do front-end.",
        "responses": {
          "200": {
            "description": "Logout realizado com sucesso. Cookies removidos."
          },
          "500": {
            "description": "Erro interno ao processar o logout."
          }
        }
      }
    },
    "/api/v1/authentication/register-system-user": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Registro de Proprietário de Sistema",
        "description": "Cria um novo usuário com a role \"Owner\" (Proprietário) vinculado a um sistema específico (`IdSystem`).\r\n\r\n**Regras de Negócio:**\r\n- O E-mail fornecido não pode existir previamente na base de dados.\r\n- O ID do sistema deve corresponder a um registro válido.\r\n- Um documento fictício único é gerado automaticamente.\r\n- Após o sucesso, um e-mail de boas-vindas com as cores do sistema é enviado.\r\n- A senha deve obedecer a política de segurança (letras, números e caracteres especiais).",
        "requestBody": {
          "description": "Dados necessários para criar a conta (Nome, E-mail, Senha, Sistema, etc).",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterSystemUserDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterSystemUserDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterSystemUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário cadastrado com sucesso."
          },
          "400": {
            "description": "Dados inválidos, e-mail já em uso, sistema inexistente ou senha fraca."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/api/v1/authentication/forgot-password": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Solicitação de Recuperação de Senha",
        "description": "Envia um e-mail contendo um link seguro para redefinição de senha, caso o e-mail exista na base.\r\n\r\n**Comportamento de Segurança:**\r\n- Para evitar enumeração de contas, a API sempre retorna 200 OK informando que \"Se o e-mail existir, o link foi enviado\", mesmo que o e-mail não seja encontrado.\r\n- O link gerado expira em 2 horas.",
        "requestBody": {
          "description": "E-mail do usuário e opcionalmente o ID do sistema para customização do template.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Requisição processada com sucesso (e-mail enviado se a conta existir)."
          },
          "500": {
            "description": "Erro interno ao gerar o token ou enviar o e-mail."
          }
        }
      }
    },
    "/api/v1/authentication/reset-password": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Redefinição de Senha",
        "description": "Processa o token gerado na recuperação de senha e define a nova senha fornecida.\r\nO token enviado pelo cliente deve estar em formato `Base64Url`.",
        "requestBody": {
          "description": "E-mail, token de recuperação e a nova senha.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Senha atualizada com sucesso."
          },
          "400": {
            "description": "Usuário não encontrado, token inválido, expirado, ou nova senha fraca."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/api/v1/authentication/change-password": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Alteração de Senha do Usuário Logado",
        "description": "Permite que um usuário autenticado mude sua própria senha.\r\nÉ obrigatório informar a senha atual para garantir a segurança da operação.",
        "requestBody": {
          "description": "Objeto contendo a senha atual e a nova senha.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Senha alterada com sucesso."
          },
          "400": {
            "description": "Senha atual incorreta ou nova senha não atende aos requisitos."
          },
          "401": {
            "description": "Usuário não autenticado (Token JWT ausente ou inválido)."
          },
          "404": {
            "description": "Registro do usuário autenticado não foi encontrado no banco de dados."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthPersonDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "nullable": true
          },
          "newPassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ForgotPasswordDto": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "idSystem": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LoginDto": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "nullable": true
          },
          "expiration": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          },
          "person": {
            "$ref": "#/components/schemas/AuthPersonDto"
          },
          "systemIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "systemNames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RegisterSystemUserDto": {
        "required": [
          "birthDate",
          "email",
          "idSystem",
          "name",
          "password",
          "phone"
        ],
        "type": "object",
        "properties": {
          "idSystem": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "password": {
            "minLength": 6,
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "phone": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ResetPasswordDto": {
        "required": [
          "email",
          "newPassword",
          "token"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "token": {
            "type": "string",
            "nullable": true
          },
          "newPassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Cole APENAS o token JWT.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}